+++ src/Ext/TechnoType/Body.cpp
@@ -161,17 +161,11 @@ void TechnoTypeExt::ExtData::LoadFromINIFile(CCINIClass* const pINI)
 	this->Ammo_DeployUnlockMaximumAmount.Read(exINI, pSection, "Ammo.DeployUnlockMaximumAmount");
 
 	this->AutoDeath_Behavior.Read(exINI, pSection, "AutoDeath.Behavior");
+	this->Convert_AutoDeath.Read(exINI, pSection, "Convert.AutoDeath");
 	this->AutoDeath_VanishAnimation.Read(exINI, pSection, "AutoDeath.VanishAnimation");
-	this->AutoDeath_OnAmmoDepletion.Read(exINI, pSection, "AutoDeath.OnAmmoDepletion");
-	this->AutoDeath_AfterDelay.Read(exINI, pSection, "AutoDeath.AfterDelay");
-	this->AutoDeath_TechnosDontExist.Read(exINI, pSection, "AutoDeath.TechnosDontExist");
-	this->AutoDeath_TechnosDontExist_Any.Read(exINI, pSection, "AutoDeath.TechnosDontExist.Any");
-	this->AutoDeath_TechnosDontExist_AllowLimboed.Read(exINI, pSection, "AutoDeath.TechnosDontExist.AllowLimboed");
-	this->AutoDeath_TechnosDontExist_Houses.Read(exINI, pSection, "AutoDeath.TechnosDontExist.Houses");
-	this->AutoDeath_TechnosExist.Read(exINI, pSection, "AutoDeath.TechnosExist");
-	this->AutoDeath_TechnosExist_Any.Read(exINI, pSection, "AutoDeath.TechnosExist.Any");
-	this->AutoDeath_TechnosExist_AllowLimboed.Read(exINI, pSection, "AutoDeath.TechnosExist.AllowLimboed");
-	this->AutoDeath_TechnosExist_Houses.Read(exINI, pSection, "AutoDeath.TechnosExist.Houses");
+
+	// AutoDeath conditions
+	ConditionGroup::ParseAutoDeath(AutoDeath_Condition, exINI, pSection);
 
 	this->Slaved_OwnerWhenMasterKilled.Read(exINI, pSection, "Slaved.OwnerWhenMasterKilled");
 	this->SlavesFreeSound.Read(exINI, pSection, "SlavesFreeSound");
@@ -515,17 +509,9 @@ void TechnoTypeExt::ExtData::Serialize(T& Stm)
 		.Process(this->Ammo_DeployUnlockMaximumAmount)
 
 		.Process(this->AutoDeath_Behavior)
+		.Process(this->Convert_AutoDeath)
 		.Process(this->AutoDeath_VanishAnimation)
-		.Process(this->AutoDeath_OnAmmoDepletion)
-		.Process(this->AutoDeath_AfterDelay)
-		.Process(this->AutoDeath_TechnosDontExist)
-		.Process(this->AutoDeath_TechnosDontExist_Any)
-		.Process(this->AutoDeath_TechnosDontExist_AllowLimboed)
-		.Process(this->AutoDeath_TechnosDontExist_Houses)
-		.Process(this->AutoDeath_TechnosExist)
-		.Process(this->AutoDeath_TechnosExist_Any)
-		.Process(this->AutoDeath_TechnosExist_AllowLimboed)
-		.Process(this->AutoDeath_TechnosExist_Houses)
+		.Process(this->AutoDeath_Condition)
 
 		.Process(this->Slaved_OwnerWhenMasterKilled)
 		.Process(this->SlavesFreeSound)