+++ src/Utilities/Enum.h
@@ -148,6 +148,13 @@ enum class AutoDeathBehavior
 	Kill = 0,     // default death option
 	Vanish = 1,
 	Sell = 2,     // buildings only
+	Convert = 3,  // units only
+};
+
+enum class ConditionGroupType
+{
+	AutoDeath = 0,
+	SW = 1,
 };
 
 enum class SelfHealGainType